Upgraded to Kotlin 1.9/JDK 17/Gradle 7.6
removed dependency on tokens-sdk, this is required because tokens-sdk has dependency on this module which causes a circular dependency loop. This has meant that the tests in this module have been moved to the tokens sdk, as they used the tokens sdk.

removed unit test support in jenkinsfile as all the tests have been moved to the tokens-sdk

Updated to latest corda os snapshot

generate a preview version for not non-release artifact publishing

modified build to use git commit id as part of preview version

updated corda os version

modified to use artfacory for corda-dev repo

hard wired version for CRAFT

wip